+++ src/app/organizations/[name]/offers/offer-container.tsx
@@ -3,6 +3,7 @@
 import { useOffers } from '@/hooks/use-offers';
 import styles from '@/styles/app/organizations/[name]/offers/page.module.scss';
 import { Offer } from '@/app/organizations/[name]/offers/offer';
+import { Loader } from '@/components/common/loader';
 
 interface OfferContainerProps {
     organizationName: string;
@@ -14,10 +15,10 @@ export function OfferContainer({ organizationName }: OfferContainerProps) {
     return (
         <>
             <h3 style={{ textAlign: 'center', margin: '2rem 0 0.5rem 0' }}>Offers</h3>
-            <div className={styles['offer-container']}>
+            <div className={styles['offer-container']} data-is-loading={isLoading}>
                 {
                     isLoading
-                        ? 'Loading offers...'
+                        ? <Loader />
                         : offers.map(offer => <Offer offer={offer} key={offer.name} />)
                 }
             </div>