+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.22.0] - 2024-05-15
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): `Loader` component.
+- [Marek Kurańda](https://github.com/mjkuranda): `Loader`s to appropriate places in app.
+- [Marek Kurańda](https://github.com/mjkuranda): `SettingsCreatorContext` to store loading state while submitting any creator form.
+
 ## [0.21.0] - 2024-05-15
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): `/manage` subpage.